requirements.txt
fastapi
uvicorn
python-multipart
firebase-admin
sendgrid
python-dotenv
openai
